+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* GIPHY integration
+* Zulip push notification integration
 * Fix full text search
 * Fix issue with NLB dependency during deletion
 * Initial development